refactor(ui): extract the perspective purge out of DataTable
AuthSessionGuard renders in the backend shell on every backend page, so importing
clearAllPerspectiveState from DataTable would have pulled the whole table
implementation into the bundle of pages that render no table. The purge only
needs the two storage prefixes, so it moves to its own module.

DataTable re-exports it, keeping `@open-mercato/ui/backend/DataTable` as the
published import path — an import path is a contract surface per
BACKWARD_COMPATIBILITY.md, and login.tsx plus the existing storage tests import
it from there.
